# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe API Docs",
        "description": "API documentation for the Recipe Application",
        "contact": {
            "name": "Recipe App Support",
            "email": "support@recipeapp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9020",
            "description": "Version 1 of the Recipe API"
        },
        {
            "url": "http://localhost:9020/"
        }
    ],
    "paths": {
...
        "/api/v1/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get a specific recipe",
                "description": "Retrieve a single recipe by ID. Only the owner can access it.",
                "operationId": "54c01aaa8806568775bb5c0064693112",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Spaghetti Carbonara"
                                        },
                                        "cuisine_type": {
                                            "type": "string",
                                            "example": "Italian"
                                        },
                                        "ingredients": {
                                            "type": "string",
                                            "example": "200g pasta\\n4 eggs\\n100g pancetta"
                                        },
                                        "steps": {
                                            "type": "string",
                                            "example": "1. Boil pasta\\n2. Fry pancetta\\n3. Mix with eggs"
                                        },
                                        "picture": {
                                            "type": "string",
                                            "example": "recipes/abc123.jpg",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-05T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-05T12:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - You do not own this recipe"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update a recipe",
                "description": "Update a recipe owned by the authenticated user. All fields are optional.",
                "operationId": "b35fa346be680db242bb9e0f01cc5067",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Partial recipe data (only fields to update)",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New name",
                                        "type": "string",
                                        "example": "Updated Carbonara"
                                    },
                                    "cuisine_type": {
                                        "description": "New cuisine",
                                        "type": "string",
                                        "example": "Italian"
                                    },
                                    "ingredients": {
                                        "description": "New ingredients",
                                        "type": "string",
                                        "example": "300g pasta\\n5 eggs\\n150g pancetta"
                                    },
                                    "steps": {
                                        "description": "New steps",
                                        "type": "string",
                                        "example": "1. Boil more pasta\\n2. Fry extra pancetta"
                                    },
                                    "picture": {
                                        "description": "Optional: Replace image (deletes old one)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Carbonara"
                                        },
                                        "cuisine_type": {
                                            "type": "string",
                                            "example": "Italian"
                                        },
                                        "ingredients": {
                                            "type": "string",
                                            "example": "300g pasta\\n5 eggs\\n150g pancetta"
                                        },
                                        "steps": {
                                            "type": "string",
                                            "example": "1. Boil more pasta\\n2. Fry extra pancetta"
                                        },
                                        "picture": {
                                            "type": "string",
                                            "example": "recipes/xyz789.jpg",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not owner"
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete a recipe",
                "description": "Permanently delete a recipe. Only the owner can delete it. The associated picture is also removed from storage.",
                "operationId": "50ca486ab0aea2c0fdb6b503837d5ed2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recipe deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - You do not own this recipe"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}